Use a single directory for compiled beam files
While it's manageable to figure out the ebin directory that rebar3
will write source files to test files are e.g. ct suite files are
problematic since rebar3 compies these files into the the _build
directory. This causes a problem:

```
rebar3 ct
```

Now the suite files are copied to the _build directory, the watcher
discovers that these files are changed we re-compile and add the path
where the beam is written to the code path, which is different for the
suite file managed by git. The consequence, watch and re-compile works
but we reload the beam file generated by rebar3 instead of the one
just re-compiled.

Having a single directory for beam files generated by erl_dev_tools
avoid all of these problems.